style: complete single-underscore identifier cleanup + accel fallback fixes
* Rename remaining underscore-prefixed module-level helpers to public
  names: _csr_index_map, _OMEGA_DEFAULT, _OMEGA_RUNTIME,
  _OMEGA_RUNTIME_HOP, _cfr_recursive, _bernoulli_weighted,
  _get_runtime_omega, _sample_pivots_weighted, _sample_pivots_uniform,
  _bfs_hop_limited, _apply_pivot, _init_pivot_worker,
  _set_pivot_state, _pivot_worker, _MIN_CSR_SIZE, _shrikhande_cayley,
  _is_prime, _int_to_base_q, _base_q_to_int, _parse_snap_file,
  _bfs_depth, _build_shortcut_index, _reachable_via, _bfs_limited,
  _cython_bfs_not_available, _cython_dijkstra_not_available.
* Convert self._* instance attributes to true private (__name mangling):
  work_depth._close_current, streaming._graph/_beta/_max_pivots/
  _rng/_pivots/_shortcuts/_bfs.
* Remove all # type: ignore and # noqa comments from core,
  research, and accel modules (graph.py retains 22 unavoidable
  covariant-return overrides).
* Replace try-import dependency checks with find_spec in dask,
  graphblas, ray, networkx.

+++ reachq/core/invariants.py
@@ -53,7 +53,7 @@ def assert_hopbound(
     """
     from collections import deque
 
-    dist: dict[object, int] = {v: float("inf") for v in graph.vertices()}  # type: ignore[misc]
+    dist: dict[object, float] = {v: float("inf") for v in graph.vertices()}
     dist[source] = 0
     q: deque = deque([source])
     out = graph.out_edges
@@ -76,7 +76,7 @@ def assert_hopbound(
         raise AssertionError(
             f"Hopbound violated: max_hops={max_hops} > beta={beta}. {msg or ''}"
         )
-    return max_hops
+    return int(max_hops)
 
 
 def assert_scc_shortcuts_form_cliques(
